# api_definition.yaml
openapi: 3.0.0
info:
  title: Quiz Master API
  description: Advanced Quiz Application API with Celery background jobs, analytics, and comprehensive quiz management
  version: 2.0.0
  contact:
    name: Quiz Master API Support
    email: support@quizmaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
paths:
...
  /api/users:
    get:
      summary: Get all users with pagination
      description: Retrieve a paginated list of all users
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: Number of users per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of users with pagination info
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/users/{user_id}:
    get:
      summary: Get user by ID
      description: Retrieve detailed information about a specific user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details with statistics
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      statistics:
                        type: object
                        properties:
                          total_attempts:
                            type: integer
                          average_score:
                            type: number
...
  /api/export/users:
    get:
      summary: Export users data
      description: Export users data in JSON or CSV format
      parameters:
        - name: format
          in: query
          description: Export format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Exported data
          content:
            application/json:
              schema:
                type: object
                properties:
                  export_type:
                    type: string
                  format:
                    type: string
                  timestamp:
                    type: string
                  data:
                    type: array
            text/csv:
              schema:
                type: string
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer